requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
diff-match-patch==20121119
Django==2.0.6
django-cors-headers==2.2.0
django-extensions==2.0.7
django-import-export==1.1.0
django-js-asset==1.1.0
django-mail-templated==2.6.5
django-mptt==0.9.0
django-pyodbc-azure==2.0.6.1
django-qr-code==1.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
jdcal==1.4
kombu==4.2.1
lxml==4.2.3
num2words==0.5.7
odfpy==1.3.6
openpyxl==2.5.8
Pillow==5.1.0
pyodbc==4.0.23
PyPDF2==1.26.0
pypng==0.0.18
PyQRCode==1.2.1
pytz==2018.4
PyYAML==3.13
qrcode==6.0
reportlab==3.4.0
requests==2.19.1
runscript==0.2.6
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
vine==1.1.4
webencodings==0.5.1
xhtml2pdf==0.2.2
xlrd==1.1.0
xlwt==1.3.0
